\name{initproject-package}
\alias{initproject-package}
\alias{initproject}
\docType{package}
\title{
initproject
}
\description{
Tools to start a new r project
}
\details{
\tabular{ll}{
Package: \tab initproject\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-02\cr
License: \tab CC0 - No Rights Reserved \cr
}
Tools to start a new r project.  Currently has only one function: establishPaths()
}
\author{
Matthew Baggott

Maintainer:  <matt@baggott.net>
}
\references{

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
}
\examples{
require(initproject)
establishPaths(projectname="fall_sales_analysis", codelocation="coding")
}
